+++ src/api/control/error_codes.rs
@@ -427,6 +427,7 @@ impl From<RoomError> for ErrorResponse {
             E::EndpointAlreadyExists(id) => {
                 Self::new(ErrorCode::EndpointAlreadyExists, &id)
             }
+            E::TryFromElement(id) => Self::new(ErrorCode::NotMemberInSpec, &id),
             E::WrongRoomId(_, _)
             | E::PeerNotFound(_)
             | E::CallbackClientError(_)