+++ src/conf/server.rs
@@ -2,6 +2,7 @@
 
 use std::net::{IpAddr, Ipv4Addr, SocketAddr};
 
+use derive_more::{Display, From};
 use serde::{Deserialize, Serialize};
 use smart_default::SmartDefault;
 
@@ -34,8 +35,8 @@ pub struct ClientApiHttpServer {
     /// [Client API]: https://tinyurl.com/yx9thsnr
     /// [Control API]: https://tinyurl.com/yxsqplq7
     /// [Jason]: https://github.com/instrumentisto/medea/tree/master/jason
-    #[default = "ws://127.0.0.1:8080/ws"]
-    pub public_url: String,
+    #[default(PublicUrl("ws://127.0.0.1:8080/ws".into()))]
+    pub public_url: PublicUrl,
 
     /// IP address to bind HTTP server to.
     ///
@@ -58,6 +59,18 @@ impl ClientApiHttpServer {
         (self.bind_ip, self.bind_port).into()
     }
 }
+
+/// Public URL of HTTP server. Address for exposed [Client API].
+/// It's assumed that HTTP server can be reached via this URL externally.
+///
+/// This address is returned from [Control API] in `sids` field and [Jason] uses
+/// this address to start its session.
+///
+/// [Client API]: https://tinyurl.com/yx9thsnr
+/// [Control API]: https://tinyurl.com/yxsqplq7
+/// [Jason]: https://github.com/instrumentisto/medea/tree/master/jason
+#[derive(Clone, Debug, Display, Deserialize, Serialize, From)]
+pub struct PublicUrl(pub String);
 
 /// [Control API] servers settings.
 ///