+++ tests/e2e/world/mod.rs
@@ -332,6 +332,49 @@ impl World {
     }
 
     /// Creates `WebRtcPublishEndpoint`s and `WebRtcPlayEndpoint`s for the
+    /// provided [`MembersPair`] using an `Apply` method of Control API.
+    pub async fn interconnect_members_via_apply(&mut self, pair: MembersPair) {
+        let mut spec = self.get_spec().await;
+        if let Some(proto::RoomElement::Member(member)) =
+            spec.pipeline.get_mut(&pair.left.id)
+        {
+            member.pipeline.insert(
+                "publish".to_string(),
+                proto::Endpoint::WebRtcPublishEndpoint(
+                    pair.left.publish_endpoint().unwrap(),
+                ),
+            );
+            let play_endpoint = pair
+                .left
+                .play_endpoint_for(&self.room_id, &pair.right)
+                .unwrap();
+            member.pipeline.insert(
+                play_endpoint.id.clone(),
+                proto::Endpoint::WebRtcPlayEndpoint(play_endpoint),
+            );
+        }
+        if let Some(proto::RoomElement::Member(member)) =
+            spec.pipeline.get_mut(&pair.right.id)
+        {
+            member.pipeline.insert(
+                "publish".to_string(),
+                proto::Endpoint::WebRtcPublishEndpoint(
+                    pair.right.publish_endpoint().unwrap(),
+                ),
+            );
+            let play_endpoint = pair
+                .right
+                .play_endpoint_for(&self.room_id, &pair.left)
+                .unwrap();
+            member.pipeline.insert(
+                play_endpoint.id.clone(),
+                proto::Endpoint::WebRtcPlayEndpoint(play_endpoint),
+            );
+        }
+        self.apply(spec).await;
+    }
+
+    /// Creates `WebRtcPublishEndpoint`s and `WebRtcPlayEndpoint`s for the
     /// provided [`MembersPair`].
     pub async fn interconnect_members(
         &mut self,
@@ -466,6 +509,31 @@ impl World {
             .into_iter()
             .filter(|i| i.fid.contains(&self.room_id))
             .collect()
+    }
+
+    /// Returns [`proto::Room`] spec of the `Room` created for this [`World`].
+    pub async fn get_spec(&mut self) -> proto::Room {
+        let el = self
+            .control_client
+            .get(&self.room_id)
+            .await
+            .unwrap()
+            .element
+            .unwrap();
+        if let proto::Element::Room(room) = el {
+            room
+        } else {
+            panic!("Returned not Room element")
+        }
+    }
+
+    /// Applies provided [`proto::Room`] spec to the `Room` created for this
+    /// [`World`].
+    pub async fn apply(&mut self, el: proto::Room) {
+        self.control_client
+            .apply(&self.room_id, proto::Element::Room(el))
+            .await
+            .unwrap();
     }
 }